docs(tt_um_6502): change table formatting
- Split large instruction table into multiple, smaller ones
- Use raw text for flag column

The original table was one large collection of instructions, which were
separated using a table-wide cell to show that the following
instructions were related to a certain group.

Unfortunately it was not possible to use markdown to retain the same
look, so I've opted to split them up instead.

Conveys the same information, but just in multiple tables now.